Refactor edit-ema-editor.component.html and edit-ema-toolbar.component.html
