+++ core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/edit-ema-toolbar/edit-ema-toolbar.component.html
@@ -1,67 +1,78 @@
-<ng-content select="[title]"></ng-content>
-<p-toolbar>
-    <div class="p-toolbar-group-left" data-testId="toolbar-left-content">
-        @if(es.previewState.editorMode === editorMode.EDIT) {
-        <p-button
-            [label]="'editpage.toolbar.preview.page' | dm"
-            (onClick)="deviceSelector.openMenu($event)"
-            icon="pi pi-desktop"
-            styleClass="p-button-text p-button-sm"
-            data-testId="ema-preview" />
-        <p-button
-            *ngIf="es.editor.urlContentMap as urlContentMap"
-            [label]="'editpage.toolbar.edit.url.map.content' | dm : [urlContentMap.contentType]"
-            (onClick)="editContentMap(urlContentMap)"
-            icon="pi pi-pencil"
-            styleClass="p-button-text p-button-sm"
-            data-testId="edit-url-content-map" />
+<ng-container *ngIf="editorState$ | async as es">
+    <dot-device-selector-seo
+        #deviceSelector
+        [apiLink]="es.iframeURL"
+        [hideSocialMedia]="!!es.clientHost"
+        (selected)="updateCurrentDevice($event)"
+        (changeSeoMedia)="onSeoMediaChange($event)"
+        appendTo="body"
+        data-testId="dot-device-selector" />
 
-        } @else {
-        <p-button
-            *ngIf="es.previewState.editorMode === editorMode.PREVIEW"
-            [label]="'editema.editor.preview.back-to-edit' | dm"
-            (onClick)="goToEditMode()"
-            icon="pi pi-arrow-left"
-            styleClass="p-button-text p-button-sm"
-            data-testId="ema-back-to-edit" />
-        }
+    <ng-content select="[title]"></ng-content>
+    <p-toolbar>
+        <div class="p-toolbar-group-left" data-testId="toolbar-left-content">
+            @if(es.previewState.editorMode === editorMode.EDIT) {
+            <p-button
+                [label]="'editpage.toolbar.preview.page' | dm"
+                (onClick)="deviceSelector.openMenu($event)"
+                icon="pi pi-desktop"
+                styleClass="p-button-text p-button-sm"
+                data-testId="ema-preview" />
+            <!-- <p-button
+                *ngIf="es.editor.urlContentMap as urlContentMap"
+                [label]="'editpage.toolbar.edit.url.map.content' | dm : [urlContentMap.contentType]"
+                (onClick)="editContentMap(urlContentMap)"
+                icon="pi pi-pencil"
+                styleClass="p-button-text p-button-sm"
+                data-testId="edit-url-content-map" /> -->
 
-        <dot-ema-bookmarks [url]="es.favoritePageURL" />
-        <p-button
-            [cdkCopyToClipboard]="es.iframeURL"
-            [label]="'editpage.header.copy' | dm"
-            (cdkCopyToClipboardCopied)="triggerCopyToast()"
-            icon="pi pi-copy"
-            styleClass="p-button-text p-button-sm"
-            data-testId="ema-copy-url" />
-        <a
-            class="p-button-sm p-button-text"
-            [href]="es.apiURL"
-            [title]="es.apiURL"
-            pButton
-            target="_blank"
-            data-testId="ema-api-link"
-            icon="pi pi-link"
-            label="API">
-        </a>
-    </div>
-    <div class="p-toolbar-group-right" data-testId="toolbar-right-content">
-        <dot-ema-running-experiment
-            [runningExperiment]="es.runningExperiment"
-            data-testId="ema-running-experiment" />
-        <dot-edit-ema-language-selector
-            [language]="es.editor.viewAs.language"
-            (selected)="onLanguageSelected($event)"
-            data-testId="language-selector" />
-        <dot-edit-ema-persona-selector
-            #personaSelector
-            [pageId]="es.editor.page.identifier"
-            [value]="es.editor.viewAs.persona"
-            (selected)="onPersonaSelected($event)"
-            (despersonalize)="onDespersonalize($event)"
-            data-testId="persona-selector" />
-        <dot-edit-ema-workflow-actions
-            [inode]="es.editor.page.inode"
-            (newPage)="handleNewPage($event)" />
-    </div>
-</p-toolbar>
+            } @else {
+            <p-button
+                *ngIf="es.previewState.editorMode === editorMode.PREVIEW"
+                [label]="'editema.editor.preview.back-to-edit' | dm"
+                (onClick)="goToEditMode()"
+                icon="pi pi-arrow-left"
+                styleClass="p-button-text p-button-sm"
+                data-testId="ema-back-to-edit" />
+            }
+
+            <dot-ema-bookmarks [url]="es.favoritePageURL" />
+            <p-button
+                [cdkCopyToClipboard]="es.iframeURL"
+                [label]="'editpage.header.copy' | dm"
+                (cdkCopyToClipboardCopied)="triggerCopyToast()"
+                icon="pi pi-copy"
+                styleClass="p-button-text p-button-sm"
+                data-testId="ema-copy-url" />
+            <a
+                class="p-button-sm p-button-text"
+                [href]="es.apiURL"
+                [title]="es.apiURL"
+                pButton
+                target="_blank"
+                data-testId="ema-api-link"
+                icon="pi pi-link"
+                label="API">
+            </a>
+        </div>
+        <div class="p-toolbar-group-right" data-testId="toolbar-right-content">
+            <dot-ema-running-experiment
+                [runningExperiment]="es.runningExperiment"
+                data-testId="ema-running-experiment" />
+            <dot-edit-ema-language-selector
+                [language]="es.editor.viewAs.language"
+                (selected)="onLanguageSelected($event)"
+                data-testId="language-selector" />
+            <dot-edit-ema-persona-selector
+                #personaSelector
+                [pageId]="es.editor.page.identifier"
+                [value]="es.editor.viewAs.persona"
+                (selected)="onPersonaSelected($event)"
+                (despersonalize)="onDespersonalize($event)"
+                data-testId="persona-selector" />
+            <dot-edit-ema-workflow-actions
+                [inode]="es.editor.page.inode"
+                (newPage)="handleNewPage($event)" />
+        </div>
+    </p-toolbar>
+</ng-container>
